# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brainiac",
    "description": "GitOps for organizational AI knowledge — capture, govern, and serve an organization's memory. Every read is RLS-scoped to the caller: an agent can never retrieve what its operator cannot.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8600",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "GET /v1/projects — the org's projects with their whitelisted repos.",
        "description": "The org's projects with their whitelisted repos (admin).",
        "operationId": "projects_list",
        "responses": {
          "200": {
            "description": "Projects of the caller's org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "POST /v1/projects — create a project.",
        "description": "Create a project (admin). Names are unique per org.",
        "operationId": "project_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Empty or oversized name"
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          },
          "409": {
            "description": "A project with this name already exists"
          }
        }
      }
    },
    "/v1/projects/{id}/repos": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "POST /v1/projects/{id}/repos — whitelist a remote under a project.",
        "description": "Whitelist a git remote under a project (admin). The remote is normalized to `host/owner/name`; within an org a remote maps to exactly one project.",
        "operationId": "repo_add",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRepoBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Repo whitelisted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddedRepoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Remote is not a recognizable git remote URL"
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          },
          "404": {
            "description": "Project not found in this org"
          },
          "409": {
            "description": "This remote is already whitelisted in the org"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddRepoBody": {
        "type": "object",
        "required": [
          "remote"
        ],
        "properties": {
          "remote": {
            "type": "string",
            "description": "Any spelling of the remote (https, ssh, scp, bare); stored normalized."
          }
        }
      },
      "AddedRepoResponse": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "remote"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "remote": {
            "type": "string",
            "description": "The normalized form that was stored — what onboarding will match on."
          }
        }
      },
...
      "CreateProjectBody": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreatedProjectResponse": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ProjectRepoView": {
        "type": "object",
        "required": [
          "id",
          "remote",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "remote": {
            "type": "string",
            "description": "Normalized remote, e.g. `github.com/acme/payments`."
          }
        }
      },
      "ProjectView": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "repos"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectRepoView"
            }
          }
        }
      },
      "ProjectsListResponse": {
        "type": "object",
        "required": [
          "projects"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectView"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Env-map token (unrestricted) or a managed `brk_…` key carrying the read/write/admin scope the route requires."
      }
    }
  },
...
}